Update version 1.0.0-beta.3

+++ src/modules/delete.php
@@ -1,5 +1,8 @@
 <?php
 
+	// Load user session
+	session_start();
+
 	// Load application files
 	require_once "../autoload.php";
 
@@ -12,14 +15,32 @@
 
 		try {
 			// Deletes the publication in the database
-			$stmt = $conn->prepare("DELETE FROM QUOTES WHERE ID_QUOTE=:id");
+			$stmt = $conn->prepare("DELETE FROM LIKES WHERE ID_QUOTE = :id");
 			$stmt->bindParam(":id", $id);
-			$stmt->execute();
+			$result = $stmt->execute();
+
+			if ($result) {
+				$stmt = $conn->prepare("DELETE FROM QUOTES WHERE ID_QUOTE = :id");
+				$stmt->bindParam(":id", $id);
+				$result = $stmt->execute();
+			}
+
+			if ($result === false) {
+				$_SESSION["message"] = "Your post cannot be deleted at this time. Please try again later.";
+			}
 
 			// Redirect to homepage
-			header('location: ../../public/index.php?page=home.php');
+			header('location: ../../public/index.php?page=home');
 		} catch (PDOException $e) {
-			header('location: ../../public/index.php?page=home.php&error=' . $e->getMessage());
+			$_SESSION["message"] = "<strong>DataBase Error</strong>: The post could not be deleted.<br>" . $e->getMessage();
+
+			// Redirect to homepage
+			header('location: ../../public/index.php?page=home');
+		} catch (Exception $e) {
+			$_SESSION["message"] = "<strong>General Error</strong>: The post could not be deleted.<br>" . $e->getMessage();
+
+			// Redirect to homepage
+			header('location: ../../public/index.php?page=home');
 		} finally {
 			// Destroy the database connection
 			$conn = null;
